Move files into features directory
